# docs/api/vedal-admin-openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: sales@vedal-med.ru
    name: VEDAL
    url: tel:88006003449
  description: |
    Контракт сайта vedal-med.ru с серверной частью портала.

    Двери и их правила:

    - `/api/public/v1/**` — чтение опубликованного. Ответы кэшируются
      на пять минут (`Cache-Control: max-age=300, public`); файлы
      документов не кэшируются вовсе, потому что снятая с публикации
      редакция не должна жить в кэшах прокси.
    - `/api/forms/v1/**` — приём заявок. Отвечает `202 Accepted`:
      заявка принята, дальше её разбирает менеджер.
    - `/api/assistant/v1/**` — вопрос ассистенту.

    Ошибки во всех дверях — `application/problem+json` по RFC 9457.
    Разбор по полям формы приезжает в расширении `fields`.

    Кросс-доменные запросы разрешены только источникам из
    `vedal.web.allowed-origins`, без передачи учётных данных: у
    публичных дверей нет ни cookie, ни сессии. Повторную отправку
    заявки различает заголовок `Idempotency-Key`, а не пользователь.

    Лимит частоты считается по адресу клиента, у каждой двери свой
    бюджет: формы — 5 обращений за 10 минут, ассистент — 20.
    Превышение отдаёт `429`.
  title: VEDAL Portal — публичное API
  version: v1
servers:
- description: Generated server url
  url: http://localhost:8081
...
paths:
...
  /api/admin/v1/clients/{id}/history:
    get:
      operationId: history_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminInteraction"
          description: OK
      security:
      - keycloak: []
      summary: История переписки и звонков по клиенту
      tags:
      - "Админка: клиенты"
    post:
      description: "История только дописывается: правки и удаления здесь нет. Исто\
        рию, которую можно поправить задним числом, незачем вести."
      operationId: addToHistory_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNewInteraction"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInteraction"
          description: Created
      security:
      - keycloak: []
      summary: Дописать в историю клиента
      tags:
      - "Админка: клиенты"
...
  /api/admin/v1/deals/{id}/history:
    get:
      operationId: history_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminInteraction"
          description: OK
      security:
      - keycloak: []
      summary: История переписки и звонков по сделке
      tags:
      - "Админка: сделки"
    post:
      description: "История только дописывается: правки и удаления здесь нет."
      operationId: addToHistory_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNewInteraction"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInteraction"
          description: Created
      security:
      - keycloak: []
      summary: Дописать в историю сделки
      tags:
      - "Админка: сделки"
...
  /api/admin/v1/leads:
    get:
      operationId: leads
      parameters:
      - description: Фильтр по статусу. Пусто — все.
        in: query
        name: status
        required: false
        schema:
          type: string
      - description: "Поиск по имени, компании, телефону и почте."
        in: query
        name: query
        required: false
        schema:
          type: string
      - description: Логин ответственного; «-» — без ответственного.
        in: query
        name: owner
        required: false
        schema:
          type: string
      - description: "Форма, с которой пришла заявка."
        in: query
        name: form
        required: false
        schema:
          type: string
      - description: Источник перехода.
        in: query
        name: source
        required: false
        schema:
          type: string
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - description: "Размер страницы, не больше 200."
        in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Page"
          description: OK
      security:
      - keycloak: []
      summary: "Заявки постранично, свежие сверху"
      tags:
      - "Админка: заявки"
  /api/admin/v1/leads/statuses:
    get:
      operationId: statuses_1
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
          description: OK
      security:
      - keycloak: []
      summary: Статусы заявки
      tags:
      - "Админка: заявки"
  /api/admin/v1/leads/{id}:
    get:
      operationId: lead
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminLead"
          description: OK
      security:
      - keycloak: []
      summary: "Заявка целиком, с текстом и версией согласия"
      tags:
      - "Админка: заявки"
  /api/admin/v1/leads/{id}/convert:
    post:
      description: |
        Заводит сделку в выбранной воронке. Клиент либо указывается существующий,
        либо заводится из данных заявки — портал не ищет совпадения сам: слить
        две карточки потом можно, разделить ошибочно слитые уже нет.

        Заявка разбирается один раз. Повтор — `409`: две сделки по одному
        обращению обе попали бы в аналитику.
      operationId: convert
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLeadConversion"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDeal"
          description: Created
      security:
      - keycloak: []
      summary: Разобрать заявку в сделку
      tags:
      - "Админка: заявки"
  /api/admin/v1/leads/{id}/history:
    get:
      operationId: history
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminInteraction"
          description: OK
      security:
      - keycloak: []
      summary: История переписки и звонков по заявке
      tags:
      - "Админка: заявки"
    post:
      description: "История только дописывается: правки и удаления здесь нет."
      operationId: addToHistory
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNewInteraction"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInteraction"
          description: Created
      security:
      - keycloak: []
      summary: Дописать в историю заявки
      tags:
      - "Админка: заявки"
...
  /api/admin/v1/leads/{id}/triage:
    post:
      description: "Изменение попадает в журнал без персональных данных — только и\
        дентификатор заявки, старый и новый статус."
      operationId: triage
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLeadTriage"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminLead"
          description: OK
      security:
      - keycloak: []
      summary: "Разобрать заявку: статус и ответственный"
      tags:
      - "Админка: заявки"
...
  /api/admin/v1/quotes/statuses:
    get:
      operationId: statuses
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
          description: OK
      security:
      - keycloak: []
      summary: Статусы КП
      tags:
      - "Админка: коммерческие предложения"
...
components:
  schemas:
...
    AdminDeal:
      type: object
      description: Карточка сделки вместе с вложениями.
      properties:
        amount:
          type: number
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AdminDealAttachment"
        clientId:
          type: string
          format: uuid
        clientName:
          type: string
        closedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        currency:
          type: string
        id:
          type: string
          format: uuid
        leadId:
          type:
          - string
          - "null"
          format: uuid
          description: "Заявка, из которой заведена сделка. Пусто у сделки, заведё\
            нной руками."
        lostReason:
          type: string
        lostStages:
          type: array
          description: "Стадии, которыми эта воронка заканчивается отказом. Перево\
            д в такую стадию требует причины проигрыша: форма спрашивает её до наж\
            атия, а не после отказа."
          items:
            type: string
        owner:
          type: string
        pipeline:
          type: string
        productSlug:
          type: string
        stage:
          type: string
        stages:
          type: array
          description: "Стадии этой воронки, по порядку. Форма рисует выбор из них\
            , а не свободное поле."
          items:
            type: string
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          description: Версия карточки. Её надо вернуть в форме правки.
        wonStages:
          type: array
          description: "Стадии, которыми эта воронка заканчивается успехом."
          items:
            type: string
...
    AdminDealAttachment:
      type: object
      description: |
        Приложенный документ. Хранится ссылка на карточку, а не копия файла:
        копия разошлась бы с оригиналом при замене ревизии.
      properties:
        attachedAt:
          type: string
          format: date-time
        attachedBy:
          type: string
        documentId:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
...
    AdminInteraction:
      type: object
      description: |
        Запись истории. Текст — персональные данные: наружу и в топики
        не уходит, туда идёт только идентификатор.
      properties:
        actor:
          type: string
        at:
          type: string
          format: date-time
        body:
          type: string
        clientId:
          type: string
          format: uuid
        dealId:
          type: string
          format: uuid
        direction:
          type: string
        id:
          type: string
          format: uuid
        kind:
          type: string
        leadId:
          type: string
          format: uuid
        subject:
          type: string
    AdminLead:
      type: object
      description: "Заявка целиком, вместе с текстом и согласием."
      properties:
        campaign:
          type:
          - string
          - "null"
          description: "Кампания, приведшая посетителя. Разрез аналитики."
        company:
          type: string
        consentAt:
          type: string
          format: date-time
        consentVersion:
          type: string
          description: "Версия текста согласия, с которым человек согласился. Хран\
            ится версия, а не галочка: через год иначе не доказать, с чем именно о\
            н согласился."
        correlationId:
          type:
          - string
          - "null"
          description: Идентификатор цепочки запроса. По нему в журнале собирается
            весь путь заявки.
        createdAt:
          type: string
          format: date-time
        dealId:
          type:
          - string
          - "null"
          format: uuid
          description: "Сделка, в которую разобрана заявка."
        email:
          type: string
        erasedAt:
          type:
          - string
          - "null"
          format: date-time
          description: "Когда персональные данные заявки уничтожены. Пусто — не ун\
            ичтожались. Нужно интерфейсу, чтобы отличить исполненное обращение от\
            \ сбоя: без этой отметки слово «удалено» в полях выглядит одинаково в\
            \ обоих случаях."
        form:
          type: string
        id:
          type: string
          format: uuid
        language:
          type:
          - string
          - "null"
          description: "Язык страницы, с которой отправлена заявка. Разрез аналити\
            ки."
          example: ru
        message:
          type: string
        name:
          type: string
        owner:
          type: string
        phone:
          type: string
        productSlug:
          type: string
        serialNumber:
          type:
          - string
          - "null"
          description: "Серийный номер изделия из сервисного обращения. Пусто — не\
            \ указан. При обезличивании не стирается: это характеристика аппарата\
            , а не человека."
        source:
          type: string
        status:
          type: string
    AdminLeadConversion:
      type: object
      description: |
        Разбор заявки в сделку. Клиент либо указывается существующий,
        либо заводится из данных заявки.
      properties:
        amount:
          type: number
          minimum: 0
        clientId:
          type:
          - string
          - "null"
          format: uuid
          description: "Существующий клиент. Пусто — карточка клиента заводится из\
            \ данных заявки. Портал не ищет совпадения сам: слить две карточки пот\
            ом можно, разделить ошибочно слитые — уже нет."
        owner:
          type: string
          maxLength: 200
          minLength: 0
        pipeline:
          type: string
          enum:
          - sales
          - dealer
          - service
          example: sales
          minLength: 1
        title:
          type:
          - string
          - "null"
          description: Название сделки. Пусто — берётся из формы и изделия заявки.
          maxLength: 300
          minLength: 0
      required:
      - pipeline
    AdminLeadTriage:
      type: object
      description: "Разбор заявки: статус и ответственный."
      properties:
        owner:
          type:
          - string
          - "null"
          description: Логин ответственного. Пусто — снять ответственного.
          maxLength: 200
          minLength: 0
        status:
          type: string
          enum:
          - draft
          - new
          - in_progress
          - won
          - lost
          minLength: 1
      required:
      - status
...
    AdminNewInteraction:
      type: object
      description: Новая запись истории.
      properties:
        at:
          type:
          - string
          - "null"
          format: date-time
          description: "Когда это было. Пусто — сейчас. Звонок записывают после ра\
            зговора, и время разговора важнее времени записи."
        body:
          type: string
          description: "Что произошло. Это персональные данные: в топики и в журна\
            л текст не уходит."
          example: "Обсудили сроки: готовы отгрузить в марте, нужен аванс 30%."
          minLength: 1
        direction:
          type:
          - string
          - "null"
          description: "Направление: `in` — от клиента, `out` — клиенту. У заметки\
            \ направления нет."
          enum:
          - in
          - out
          pattern: ^$|in|out
        kind:
          type: string
          enum:
          - call
          - email
          - meeting
          - note
          example: call
          minLength: 1
          pattern: call|email|meeting|note
        subject:
          type: string
          example: Созвон по срокам поставки
          maxLength: 300
          minLength: 0
      required:
      - body
      - kind
...
    AdminQuoteRow:
      type: object
      description: Строка списка КП.
      properties:
        createdAt:
          type: string
          format: date-time
        currency:
          type: string
        dealId:
          type: string
          format: uuid
        dealTitle:
          type: string
        id:
          type: string
          format: uuid
        number:
          type: string
        sentAt:
          type: string
          format: date-time
        status:
          type: string
        total:
          type: number
        validUntil:
          type: string
          format: date
...
    Page:
      type: object
      description: Страница списка.
      properties:
        items:
          type: array
          description: Строки страницы.
          items:
            $ref: "#/components/schemas/AdminQuoteRow"
        page:
          type: integer
          format: int32
          description: "Номер страницы, с нуля."
          example: 0
        pages:
          type: integer
          format: int32
          description: Всего страниц.
          example: 3
        size:
          type: integer
          format: int32
          description: Размер страницы.
          example: 50
        total:
          type: integer
          format: int64
          description: Всего строк во всей выборке.
          example: 137
...
  securitySchemes:
    keycloak:
      bearerFormat: JWT
      description: |
        Токен доступа из Keycloak. Роли портала — `portal-admin`,
        `portal-sales` и `portal-production` в `realm_access.roles`; токен без них
        проходит проверку подписи и получает `403`.

        В режиме `vedal.iam.mode=local` вместо этого работает
        HTTP Basic поверх учётных записей портала — режим
        разработки, не контракт.
      scheme: bearer
      type: http